+++ backend/src/main/java/io/blog/devlog/domain/file/repository/TempFileRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TempFileRepository extends JpaRepository<TempFile, Integer> {
+public interface TempFileRepository extends JpaRepository<TempFile, Long> {
 }
